# server/docs/swagger.yaml
definitions:
...
  api.AdminUserCreateReq:
    properties:
      avatar:
        description: 头像
        type: string
      email:
        description: 邮箱
        type: string
      name:
        description: 姓名
        type: string
      password:
        description: 密码
        type: string
      username:
        description: 用户名
        type: string
    required:
    - email
    - name
    - password
    - username
    type: object
  api.AdminUserCreateRes:
    properties:
      id:
        description: 管理员ID
        type: string
    type: object
  api.AdminUserDeleteReq:
    properties:
      id:
        description: ID
        type: integer
    required:
    - id
    type: object
...
  api.AdminUserListReq:
    properties:
      current:
        minimum: 1
        type: integer
      keyword:
        description: 查询关键字
        type: string
      order_key:
        type: string
      order_type:
        enum:
        - ase
        - desc
        type: string
      page_size:
        minimum: 0
        type: integer
    type: object
  api.AdminUserListRes:
    properties:
      list:
        items:
          $ref: '#/definitions/model.AdminUser'
        type: array
      total:
        type: integer
    type: object
  api.AdminUserLoginReq:
    properties:
      password:
        description: 密码
        type: string
      username:
        description: 用户名
        type: string
    type: object
  api.AdminUserLoginRes:
    properties:
      token:
        type: string
    type: object
  api.AdminUserResetPasswordReq:
    properties:
      id:
        description: ID
        type: integer
      password:
        description: 密码
        type: string
    required:
    - id
    - password
    type: object
  api.AdminUserUpdateReq:
    properties:
      avatar:
        description: 头像
        type: string
      id:
        description: ID
        type: integer
      name:
        description: 姓名
        type: string
    required:
    - id
    - name
    type: object
  api.AdminUserUpdateRoleReq:
    properties:
      role_ids:
        description: 角色 ID
        items:
          type: integer
        type: array
      user_id:
        description: 用户 ID
        type: integer
    required:
    - role_ids
    - user_id
    type: object
  api.AdminUserUpdateStatusReq:
    properties:
      id:
        description: ID
        type: integer
      status:
        allOf:
        - $ref: '#/definitions/consts.AdminUserStatus'
        description: 状态 1-解封 2-封禁
        enum:
        - -1
        - 1
    required:
    - id
    - status
    type: object
...
  consts.AdminUserStatus:
    enum:
    - 1
    - -1
    type: integer
    x-enum-comments:
      AdminUserStatusLocked: 锁定
      AdminUserStatusNormal: 正常
    x-enum-varnames:
    - AdminUserStatusNormal
    - AdminUserStatusLocked
...
  controller.loginBodyDto:
    properties:
      password:
        description: 密码
        type: string
      username:
        description: 用户名
        type: string
    type: object
  controller.loginSuccessResponse:
    properties:
      token:
        type: string
    type: object
  controller.registerBodyDto:
    properties:
      captcha:
        description: 邮箱验证码
        type: string
      email:
        description: 邮箱
        type: string
      password:
        description: 密码
        type: string
      username:
        description: 用户名
        type: string
    required:
    - email
    - password
    - username
    type: object
  controller.wxMpLoginBodyDto:
    properties:
      code:
        description: code
        type: string
    required:
    - code
    type: object
  model.AdminRole:
    properties:
      code:
        description: 角色编号
        type: string
      created_at:
        description: 创建时间
        type: string
      desc:
        type: string
      id:
        description: ID
        type: integer
      name:
        type: string
      permission_codes:
        description: 权限码
        items:
          type: string
        type: array
      updated_at:
        description: 更新时间
        type: string
      users:
        items:
          $ref: '#/definitions/model.AdminUser'
        type: array
    required:
    - code
    - created_at
    - id
    - name
    - users
    type: object
  model.AdminUser:
    properties:
      avatar:
        type: string
      created_at:
        description: 创建时间
        type: string
      email:
        type: string
      id:
        description: ID
        type: integer
      is_root:
        description: 是否是超级管理员
        type: boolean
      name:
        type: string
      news:
        items:
          $ref: '#/definitions/model.News'
        type: array
      roles:
        items:
          $ref: '#/definitions/model.AdminRole'
        type: array
      status:
        $ref: '#/definitions/consts.AdminUserStatus'
      updated_at:
        description: 更新时间
        type: string
      username:
        type: string
    required:
    - created_at
    - id
    - name
    - status
    - username
    type: object
  model.News:
    properties:
      author_id:
        description: 作者
        type: integer
      content:
        description: 内容
        type: string
      cover:
        description: 封面
        type: string
      created_at:
        description: 创建时间
        type: string
      desc:
        description: 简介
        type: string
      id:
        description: ID
        type: integer
      is_visible:
        description: 可见性
        type: boolean
      push_date:
        description: 发布日期
        type: string
      recommend:
        description: 推荐等级
        type: integer
      title:
        description: 新闻标题
        type: string
      updated_at:
        description: 更新时间
        type: string
    required:
    - author_id
    - created_at
    - id
    - is_visible
    - recommend
    - title
    type: object
...
  resp.Result:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
  title: 男生自用 API 接口文档
  version: "1.0"
paths:
...
  /api/admin/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/api.AdminUserLoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: resp
          schema:
            allOf:
            - $ref: '#/definitions/resp.Result'
            - properties:
                data:
                  $ref: '#/definitions/api.AdminUserLoginRes'
              type: object
      summary: 用户登录
      tags:
      - 管理后台-用户认证
...
  /api/admin/user/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/api.AdminUserCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: resp
          schema:
            allOf:
            - $ref: '#/definitions/resp.Result'
            - properties:
                data:
                  $ref: '#/definitions/api.AdminUserCreateRes'
              type: object
      summary: 新增管理员
      tags:
      - 管理后台-管理员用户
...
  /api/admin/user/delete:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/api.AdminUserDeleteReq'
      produces:
      - application/json
      responses:
        "200":
          description: resp
          schema:
            $ref: '#/definitions/resp.Result'
      summary: 删除管理员
      tags:
      - 管理后台-管理员用户
  /api/admin/user/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/api.AdminUserListReq'
      produces:
      - application/json
      responses:
        "200":
          description: resp
          schema:
            allOf:
            - $ref: '#/definitions/resp.Result'
            - properties:
                data:
                  $ref: '#/definitions/api.AdminUserListRes'
              type: object
      summary: 管理员列表
      tags:
      - 管理后台-管理员用户
  /api/admin/user/reset-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/api.AdminUserResetPasswordReq'
      produces:
      - application/json
      responses:
        "200":
          description: resp
          schema:
            $ref: '#/definitions/resp.Result'
      summary: 重置管理员密码
      tags:
      - 管理后台-管理员用户
  /api/admin/user/update/info:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/api.AdminUserUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: resp
          schema:
            $ref: '#/definitions/resp.Result'
      summary: 修改管理员信息
      tags:
      - 管理后台-管理员用户
  /api/admin/user/update/role:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/api.AdminUserUpdateRoleReq'
      produces:
      - application/json
      responses:
        "200":
          description: resp
          schema:
            $ref: '#/definitions/resp.Result'
      summary: 为管理用户更新角色
      tags:
      - 管理后台-管理员用户
  /api/admin/user/update/status:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/api.AdminUserUpdateStatusReq'
      produces:
      - application/json
      responses:
        "200":
          description: resp
          schema:
            $ref: '#/definitions/resp.Result'
      summary: 修改管理员状态(封禁/解封)
      tags:
      - 管理后台-管理员用户
  /api/admin/user/upload:
    post:
      consumes:
      - application/json
      parameters:
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 文件地址
          schema:
            allOf:
            - $ref: '#/definitions/resp.Result'
            - properties:
                data:
                  type: string
              type: object
      summary: 文件上传
      tags:
      - 管理后台-通用接口
  /api/admin/user/upload-oss:
    post:
      consumes:
      - application/json
      parameters:
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      - description: 文件路径
        in: formData
        name: prefix
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 文件地址
          schema:
            allOf:
            - $ref: '#/definitions/resp.Result'
            - properties:
                data:
                  type: string
              type: object
      summary: 文件上传至 阿里云 oss
      tags:
      - 管理后台-通用接口
  /api/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/controller.loginBodyDto'
      produces:
      - application/json
      responses:
        "200":
          description: resp
          schema:
            allOf:
            - $ref: '#/definitions/resp.Result'
            - properties:
                data:
                  $ref: '#/definitions/controller.loginSuccessResponse'
              type: object
      summary: 用户登录
      tags:
      - 前台-用户
  /api/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/controller.registerBodyDto'
      produces:
      - application/json
      responses:
        "200":
          description: resp
          schema:
            allOf:
            - $ref: '#/definitions/resp.Result'
            - properties:
                data:
                  $ref: '#/definitions/controller.loginSuccessResponse'
              type: object
      summary: 用户注册
      tags:
      - 前台-用户
  /api/auth/wxmp/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: body
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/controller.wxMpLoginBodyDto'
      produces:
      - application/json
      responses:
        "200":
          description: resp
          schema:
            allOf:
            - $ref: '#/definitions/resp.Result'
            - properties:
                data:
                  $ref: '#/definitions/controller.loginSuccessResponse'
              type: object
      summary: 微信小程序登录
      tags:
      - 前台-用户
...
swagger: "2.0"